test(agentic-ai): run the agent e2e suite over the native v2 providers
Enable the v1->v2 provider-config rewrite for the e2e suite (removing the
earlier switch-off pin) so v1 provider configurations execute through the
native v2 chat-model providers end to end.

Migrate the WireMock stubs to the native wire formats: OpenAI Completions
streams SSE (folding in the former V2 SSE stub, now removed and its users
repointed), Anthropic streams SSE messages, and Bedrock Converse streams the
AWS event stream. Repoint the provider wire-format fixtures and the
Azure/Foundry unified path accordingly.

Adjust assertions for the native wire shapes, which differ cosmetically from
LangChain4j but are semantically equivalent: schema-less JSON emits
json_object; strict structured output adds strict:true and
additionalProperties:false; the resolved provider is reported as "openai";
tool input schemas pass through as authored (no injected empty items or
required); user messages are encoded as content-part arrays.

The e2e suite no longer exercises the LangChain4j implementations (their
factories remain covered by unit tests); ProviderWireFormatSmokeTests still
assert that v1 configurations route to the native wire format.
