+++ connectors-e2e-test/connectors-e2e-test-agentic-ai/src/test/java/io/camunda/connector/e2e/agenticai/aiagent/task/AgentTaskAgentInstanceTests.java
@@ -101,9 +101,9 @@ class AgentTaskAgentInstanceTests extends BaseAgentTaskTest {
     AgentInstanceEngineVerifier.verify(camundaClient, agentInstanceKey.get())
         .hasStatus(AgentInstanceStatus.COMPLETED)
         .hasMetrics(new AgentMetrics(2, new AgentMetrics.TokenUsage(25, 45), 1))
-        .hasDefinition("gpt-4o", "openaiCompatible")
+        .hasDefinition("gpt-4o", "openai")
         .hasToolsContaining("SuperfluxProduct")
-        .createdWithConfigurationItem("gpt-4o", "openaiCompatible", 10)
+        .createdWithConfigurationItem("gpt-4o", "openai", 10)
         .hasConfigurationItemsAtLeast(2)
         .hasConversationRoles(
             AgentInstanceHistoryRole.USER,
@@ -182,9 +182,9 @@ class AgentTaskAgentInstanceTests extends BaseAgentTaskTest {
     AgentInstanceEngineVerifier.verify(camundaClient, agentInstanceKey.get())
         .hasStatus(AgentInstanceStatus.COMPLETED)
         .hasMetrics(new AgentMetrics(3, new AgentMetrics.TokenUsage(35, 65), 2))
-        .hasDefinition("gpt-4o", "openaiCompatible")
+        .hasDefinition("gpt-4o", "openai")
         .hasToolsContaining("SuperfluxProduct")
-        .createdWithConfigurationItem("gpt-4o", "openaiCompatible", 10)
+        .createdWithConfigurationItem("gpt-4o", "openai", 10)
         .hasConfigurationItemsAtLeast(2)
         .hasConversationRoles(
             AgentInstanceHistoryRole.USER,
