+++ connectors-e2e-test/connectors-e2e-test-agentic-ai/src/test/java/io/camunda/connector/e2e/agenticai/aiagent/task/AgentTaskResponseHandlingTests.java
@@ -145,7 +145,7 @@ public class AgentTaskResponseHandlingTests extends BaseAgentTaskTest {
                     "name",
                     expectedJsonSchemaName,
                     "strict",
-                    false,
+                    true,
                     "schema",
                     Map.of(
                         "type",
@@ -155,7 +155,9 @@ public class AgentTaskResponseHandlingTests extends BaseAgentTaskTest {
                             "text", Map.of("type", "string"),
                             "length", Map.of("type", "number")),
                         "required",
-                        List.of("text", "length"))));
+                        List.of("text", "length"),
+                        "additionalProperties",
+                        false)));
       }
     }
