+++ connectors-e2e-test/connectors-e2e-test-agentic-ai/src/test/java/io/camunda/connector/e2e/agenticai/aiagent/wiremock/anthropic/AnthropicMessagesV1WireFormatFixture.java
@@ -19,13 +19,26 @@ package io.camunda.connector.e2e.agenticai.aiagent.wiremock.anthropic;
 import com.github.tomakehurst.wiremock.junit5.WireMockRuntimeInfo;
 import io.camunda.connector.e2e.ElementTemplate;
 import io.camunda.connector.e2e.agenticai.aiagent.wiremock.spi.ProviderWireFormatFixture;
+import io.camunda.connector.e2e.agenticai.aiagent.wiremock.spi.TurnStub;
 import java.util.function.Function;
 
 /**
  * Plugs Anthropic's Messages API wire format into the provider-agnostic {@link
- * ProviderWireFormatFixture} SPI, driving the connector through the v1 element template (a plain,
- * non-streaming POST). See {@link AbstractAnthropicMessagesWireFormatFixture} for the wire-format
- * plumbing shared with the v2 fixture ({@link AnthropicMessagesV2WireFormatFixture}).
+ * ProviderWireFormatFixture} SPI, driving the connector through the v1 element template. With the
+ * v1&rarr;v2 provider-config rewrite switch on (the default), the v1 template's {@code
+ * provider.anthropic.*} config is rewritten onto the native v2 Anthropic provider at the connector
+ * boundary, so - just like {@link AnthropicMessagesV2WireFormatFixture} - the connector always
+ * calls the vendor SDK's streaming endpoint ({@code client.messages().createStreaming(params)}),
+ * which expects a real {@code text/event-stream} SSE body. {@link #stubConversation(TurnStub...)}
+ * is therefore overridden here to stub that via {@link
+ * StreamingAnthropicMessagesSseChatModelStubs}, exactly like the v2 fixture, rather than inheriting
+ * {@link AbstractAnthropicMessagesWireFormatFixture}'s buffered-JSON default (which matches the
+ * pre-rewrite v1 client and remains in use by unit tests that disable the rewrite switch).
+ *
+ * <p>See {@link AbstractAnthropicMessagesWireFormatFixture} for the *request* wire-format plumbing
+ * shared with the v2 fixture ({@code recordedRequests()}, {@code
+ * assertResponseFormatConfigured(...)}) - identical between v1 and v2 since only the response
+ * framing differs.
  */
 public final class AnthropicMessagesV1WireFormatFixture
     extends AbstractAnthropicMessagesWireFormatFixture {
@@ -33,6 +46,11 @@ public final class AnthropicMessagesV1WireFormatFixture
   @Override
   public String apiName() {
     return "AnthropicMessagesV1";
+  }
+
+  @Override
+  public void stubConversation(TurnStub... turns) {
+    StreamingAnthropicMessagesSseChatModelStubs.stubConversation(turns);
   }
 
   @Override
