+++ connectors-e2e-test/connectors-e2e-test-agentic-ai/src/test/java/io/camunda/connector/e2e/agenticai/aiagent/wiremock/openai/AzureOpenAiCompletionsWireFormatFixture.java
@@ -29,28 +29,40 @@ import java.util.function.Function;
 
 /**
  * Plugs Azure OpenAI's Chat Completions wire format into the provider-agnostic {@link
- * ProviderWireFormatFixture} SPI.
+ * ProviderWireFormatFixture} SPI, driving the connector through the v1 {@code azureOpenAi} element
+ * template. With the v1&rarr;v2 provider-config rewrite switch on (the default), the v1 template's
+ * {@code provider.azureOpenAi.*} config is rewritten onto the native v2 OpenAI provider's {@code
+ * foundry} backend ({@code OpenAiChatModelConfiguration.OpenAiBackend.OpenAiFoundryBackend} - see
+ * {@code V1ToV2ProviderConfigurationMapperImpl#mapAzureOpenAi}), which:
  *
- * <p>The request/response body shape is byte-for-byte identical to OpenAI's, so this fixture reuses
- * {@link OpenAiCompletionsChatModelStubs}, {@link OpenAiCompletionsRecordedConversation} and {@link
- * OpenAiCompletionsRecordedChatRequestAdapter} directly rather than duplicating them — only the URL
- * (deployment-based, {@code {endpoint}/openai/deployments/{deploymentId}/chat/completions}) and
- * authentication differ, which this class configures below.
- *
- * <p>Azure's SDK ({@code azure-core}'s {@code KeyCredentialPolicy}) unconditionally rejects
- * non-HTTPS endpoints when using API-key authentication (hardcoded, no builder-level bypass) — so
- * unlike the other three fixtures, this one points at WireMock's HTTPS port, whose self-signed
- * certificate ({@code BaseAgentTest.httpsKeystoreFile()}) is also configured as the JVM's trust
- * store for this test run (see {@code ProviderWireFormatSmokeTests}).
- *
- * <p>Separately, {@code langchain4j-azure-openai}'s message mapper only handles {@code
- * TextContent}/{@code ImageContent} in user messages, not {@code PdfFileContent}, unlike
- * OpenAI/Anthropic/Bedrock.
+ * <ul>
+ *   <li>always drives the vendor SDK's streaming endpoint ({@code
+ *       client.chat().completions().createStreaming(params)}) - so the response body must be a real
+ *       {@code text/event-stream} SSE stream, exactly like the plain OpenAI/{@code
+ *       openaiCompatible} rows. The request/response chunk shape is otherwise byte-for-byte
+ *       identical to OpenAI's, so this fixture reuses {@link OpenAiCompletionsChatModelStubs}
+ *       (already SSE-based), {@link OpenAiCompletionsRecordedConversation} and {@link
+ *       OpenAiCompletionsRecordedChatRequestAdapter} directly rather than duplicating them.
+ *   <li>normalizes the configured endpoint by appending a unified {@code /openai/v1} path segment
+ *       ({@code AzureUrlPathMode.UNIFIED}) rather than the deployment-based path the old
+ *       LangChain4j/azure-core client used ({@code
+ *       /openai/deployments/{deploymentId}/chat/completions}) - the openai-java SDK then appends
+ *       the family path itself, giving {@code {endpoint}/openai/v1/chat/completions}. There is no
+ *       {@code api-version} query parameter unless one is explicitly configured (the v1&rarr;v2
+ *       mapper never sets one), and the model is addressed via the body's {@code model} field, not
+ *       a deployment-name path segment.
+ *   <li>sends API-key authentication as a dedicated {@code api-key} header via the openai-java
+ *       SDK's own credential type, not {@code Authorization: Bearer} and not {@code azure-core}'s
+ *       {@code KeyCredentialPolicy} - so, unlike the pre-rewrite client, HTTPS is no longer
+ *       enforced at the credential layer. This fixture still points at WireMock's HTTPS port for
+ *       parity with real Azure endpoints; its self-signed certificate ({@code
+ *       BaseAgentTest.httpsKeystoreFile()}) is trusted JVM-wide for this test run (see {@code
+ *       ProviderWireFormatSmokeTests}).
+ * </ul>
  */
 public final class AzureOpenAiCompletionsWireFormatFixture implements ProviderWireFormatFixture {
 
-  private static final String CHAT_COMPLETIONS_PATH =
-      "/openai/deployments/test-model/chat/completions";
+  private static final String CHAT_COMPLETIONS_PATH = "/openai/v1/chat/completions";
 
   @Override
   public String apiName() {
